% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream.R
\name{Clustering.All.Rate}
\alias{Clustering.All.Rate}
\title{Conduct cell clustering on mutiple datasets and removal rates}
\usage{
Clustering.All.Rate(data.removal.all.rate)
}
\arguments{
\item{data.removal.all.rate}{A list of scRNA-seq data matrices after removing doublets by different methods with different removal rates.}
}
\value{
A list of clustering result (number of clusters) under different removal rates.
}
\description{
Conduct Louvin Clustering on datasets generated with different doublet removal rates and doublet-detection methods.
}
\examples{
result.cluster.all.rate <- Clustering.All.Rate(data.removal.all.rate)

}
