% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream.R
\name{Clustering.All}
\alias{Clustering.All}
\title{Conduct cell clustering on mutiple datasets}
\usage{
Clustering.All(count.list.all)
}
\arguments{
\item{count.list.all}{A list of scRNA-seq data matrices after removing doublets by different methods with single removal rate.}
}
\value{
A list of clustering result (number of clusters) under different removal rates.
}
\description{
Conduct Louvin Clustering on datasets generated by doublet-detection methods.
}
\examples{
result.cluster.all <- Clustering.All(count.list.all = data.removal.all$count)

}
