% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream.R
\name{Clustering.Quality}
\alias{Clustering.Quality}
\title{Calculate singlet rates}
\usage{
Clustering.Quality(
  table.cluster,
  result.cluster.all.rate,
  data.removal.all.rate
)
}
\arguments{
\item{table.cluster}{A data frame with clustering result.}

\item{result.cluster.all.rate}{A list of clustering result (number of clusters) under different removal rates.}

\item{data.removal.all.rate}{A list of scRNA-seq data matrices after removing doublets by different methods with different removal rates.}
}
\value{
A data frame with singlet rates in each correctly identified cluster.
}
\description{
Calculate singlet rates in correctly identified clusters for different doublet-detection methods.
}
\examples{
table.cluster.quality <- Clustering.Quality(table.cluster, result.cluster.all.rate, data.removal.all.rate)

}
