% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream.R
\name{Clustering}
\alias{Clustering}
\title{Conduct Louvin Clustering on single dataset}
\usage{
Clustering(count)
}
\arguments{
\item{count}{}
}
\value{
result.cluster
}
\description{
Conduct Louvin Clustering on single dataset
}
\examples{

}
