% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindDoublets.R
\name{FindACC.All}
\alias{FindACC.All}
\title{Calculate precision, recall, or true negative rate (TNR) of identified doublets on multiple datasets.}
\usage{
FindACC.All(doublet.list.all, label.list, type)
}
\arguments{
\item{doublet.list.all}{A list of doublet indices for different doublet-detection methods on multiple datasets.}

\item{label.list}{A list of vectors of 0/1 doublet annotations on multiple datasets.}
}
\value{
A list of precision, recall, or TNR for different doublet-detection methods on multiple datasets.
}
\description{
Calculate precision, recall, or TNR based on identified doublet indices and true doublet annotations on multiple datasets.
}
\examples{
precision.list.all <- FindACC.All(doublet.list.all = doublet.list.all, label.list = label.list, type = 'precision')
recall.list.all <- FindACC.All(doublet.list.all = doublet.list.all, label.list = label.list, type = 'recall')
tnr.list.all <- FindACC.All(doublet.list.all = doublet.list.all, label.list = label.list, type = 'TNR')

}
