% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindDoublets.R
\name{FindACC}
\alias{FindACC}
\title{Calculate precision, recall, or true negative rate (TNR) of identified doublets}
\usage{
FindACC(doublet.list, label, type)
}
\arguments{
\item{doublet.list}{A list of doublet indices for different doublet-detection methods.}

\item{label}{A vector of 0/1 doublet annotations.}

\item{type}{A character of "precision", "recall", or "TNR".}
}
\value{
A list of precision, recall, or TNR for different doublet-detection methods.
}
\description{
Calculate precision, recall, or TNR based on identified doublet indices and true doublet annotations.
}
\examples{
acc.list <- FindACC(doublet.list = doublet.list, label = label.list$`J293t-dm`, type = 'precision')

}
