% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindDoublets.R
\name{FindAUC.All}
\alias{FindAUC.All}
\title{Calculate AUPRC or AUROC for mutiple datasets}
\usage{
FindAUC.All(score.list.all, label.list, type)
}
\arguments{
\item{score.list.all}{A list of doublet scores on multiple datasets.}

\item{label.list}{A list of vectors of 0/1 doublet annotations.}

\item{type}{A character of "AUPRC" or "AUROC".}
}
\value{
A list of AUPRCs or AUROCs of different doublet-detection methods on multiple datasets.
}
\description{
Calculate AUPRC and AUROC based on doublet scores and annotations for different doublet-detection methods on multiple datasets.
}
\examples{
auprc.list.all <- FindAUC.All(score.list.all, label.list, 'AUPRC')

}
