% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindDoublets.R
\name{FindAUC}
\alias{FindAUC}
\title{Calculate AUPRC or AUROC for one dataset}
\usage{
FindAUC(score.list, label, type)
}
\arguments{
\item{score.list}{A list of doublet scores on one dataset.}

\item{label}{A vector of 0/1 doublet annotations.}

\item{type}{A character of "AUPRC" or "AUROC".}
}
\value{
A list of AUPRCs or AUROCs of different doublet-detection methods.
}
\description{
Calculate AUPRC and AUROC based on doublet scores and annotations for different doublet-detection methods on one dataset.
}
\examples{
auprc.list <- FindAUC(score.list = score.list, label = label.list$`J293t-dm`, type = 'AUPRC')
auroc.list <- FindAUC(score.list = score.list, label = label.list$`J293t-dm`, type = 'AUROC')

}
