% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindDoublets.R
\name{FindAUC.Single}
\alias{FindAUC.Single}
\title{Calculate AUPRC or AUROC}
\usage{
FindAUC.Single(score, label, type)
}
\arguments{
\item{score}{A vector of doublet scores.}

\item{label}{A vector of doublet annotations (0/1).}

\item{type}{A string of "AUPRC" or "AUROC".}
}
\value{
A number of AUPRC or AUROC.
}
\description{
Calculate AUPRC and AUROC based on doublet score and annotation on single dataset.
}
