% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream.R
\name{FindDE}
\alias{FindDE}
\title{Find DE genes}
\usage{
FindDE(data.removal.list, method.de = "wilcox")
}
\arguments{
\item{data.removal.list}{A list of scRNA-seq datasets with cell type annotations.}

\item{method.de}{A name of DE method: "wilcox", "bimod", "t", "poisson", "negbinom", "LR", or "MAST".}
}
\value{
A list of DE gene indices identified for each doublet-detection method.
}
\description{
Conduct DE gene analysis on a list of datasets by user-specified methods.
}
\examples{
table.DE.all <- data.frame()
for(DE.method in c('MAST', 'wilcox', 'bimod')){
 DE.list <- FindDE(data.removal.list = data.removal.list, method.de = DE.method)
 DE.acc.list <- FindDEACC(DE.list=DE.list, DE.truth=data.de$gene.de, gene.all=rownames(data.de$count))
 table.DE <- ListToDataframe(l = DE.acc.list, type = 'barplot')
 table.DE[['DE_method']] <- DE.method
 table.DE.all <- rbind(table.DE.all, table.DE)
}

}
