% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream.R
\name{FindDEACC}
\alias{FindDEACC}
\title{Calculate precision, recall, and TNR of identified DE genes}
\usage{
FindDEACC(DE.list, DE.truth, gene.all)
}
\arguments{
\item{DE.list}{A list of identified DE genes.}

\item{DE.truth}{A list of ground-truth DE genes.}

\item{gene.all}{A vector of all gene names.}
}
\value{
A list of precision, recall, and TNR across different doublet-detection methods and DE methods.
}
\description{
Calculate precision, recall, and TNR of identified DE genes across different doublet-detection methods and DE methods.
}
\examples{
table.DE.all <- data.frame()
for(DE.method in c('MAST', 'wilcox', 'bimod')){
 DE.list <- FindDE(data.removal.list = data.removal.list, method.de = DE.method)
 DE.acc.list <- FindDEACC(DE.list=DE.list, DE.truth=data.de$gene.de, gene.all=rownames(data.de$count))
 table.DE <- ListToDataframe(l = DE.acc.list, type = 'barplot')
 table.DE[['DE_method']] <- DE.method
 table.DE.all <- rbind(table.DE.all, table.DE)
}

}
