% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream.R
\name{FindDistributedAUC.All}
\alias{FindDistributedAUC.All}
\title{Calculate detection accuracy under distributed computing}
\usage{
FindDistributedAUC.All(count, label, method, batches, type)
}
\arguments{
\item{count}{A scRNA-seq data matrix.}

\item{label}{A vector of doublet annotations (0/1).}

\item{batches}{A vector of batch numbers.}

\item{type}{A string of measurement "AUPRC" or "AUROC".}
}
\value{
A list of measurement across different batch numbers and methods.
}
\description{
Randomly split one dataset into user-specified batches. Perform doublet detection on each batch by different methods to calculate doublet scores.
Concatenate doublet scores per batch. Calclulate AUPRC or AUROC on concatenated doublet scores.
}
\examples{
data.list <- ReadData(path = ".../real_datasets")
count <- data.list$count$`pbmc-ch`
label <- data.list$label$`pbmc-ch`
label <- ifelse(label == 'doublet', 1, 0)
auc.list.batch <- FindDistributedAUC.All(count = count, label = label, method=methods, batches = 2:10, type='AUPRC')

}
