% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream.R
\name{FindDistributedAUC}
\alias{FindDistributedAUC}
\title{Calculate concatenated AUPRC or AUROC}
\usage{
FindDistributedAUC(score.list.all, label.list, type)
}
\arguments{
\item{score.list.all}{A list of doublet scores of each batch.}

\item{label.list}{A list of doublet annotations of each batch.}

\item{type}{A measurement type of "AUPRC" or "AUROC".}
}
\value{
A list of AUPRC or AUROC of different batches.
}
\description{
Concatenate doublet scores of different batches. Calculate AUPRC or AUROC on concatenated doublet scores.
}
\examples{

}
