% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindDoublets.R
\name{FindDoublets.All.Rate}
\alias{FindDoublets.All.Rate}
\title{Call doublets on multiple datasets and doublet rates}
\usage{
FindDoublets.All.Rate(score.list.all, rates)
}
\arguments{
\item{score.list.all}{A list of doublet scores on multiple dataset.}

\item{rates}{A user-specified vector of doublet rates (from 0 to 1).}
}
\value{
A list of doublet indices for different doublet-detection methods on multiple datasets and doublet rates.
}
\description{
Call doublets based on doublet scores and user-specified doublet rates on multiple datasets.
}
\examples{
doublet.list.all.rate <- FindDoublets.All.Rate(score.list.all = score.list.all, rates = seq(0.01, 0.25, 0.01))

}
