% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindDoublets.R
\name{FindDoublets.All}
\alias{FindDoublets.All}
\title{Call doublets on multiple datasets}
\usage{
FindDoublets.All(score.list.all, rate)
}
\arguments{
\item{score.list.all}{A list of doublet scores on multiple dataset.}

\item{rate}{A user-specified doublet rate (from 0 to 1).}
}
\value{
A list of doublet indices for different doublet-detection methods on multiple datasets.
}
\description{
Call doublets based on doublet scores and a user-specified doublet rate on multiple datasets.
}
\examples{
doublet.list.all <- FindDoublets.All(score.list.all = score.list.all, rate = .1)

}
