% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindDoublets.R
\name{FindDoublets}
\alias{FindDoublets}
\title{Call doublets on one dataset}
\usage{
FindDoublets(score.list, rate)
}
\arguments{
\item{score.list}{A list of doublet scores on one dataset.}

\item{rate}{A user-specified doublet rate (from 0 to 1).}
}
\value{
A list of doublet indices for different doublet-detection methods.
}
\description{
Call doublets based on doublet scores and a user-specified doublet rate.
}
\examples{
doublet.list <- FindDoublets(score.list = score.list, rate = .1)

}
