% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream.R
\name{FindTempDE.All}
\alias{FindTempDE.All}
\title{Infer temporally DE genes and calculate corresponding DE genes for multiple methods}
\usage{
FindTempDE.All(data.list, gene.de)
}
\arguments{
\item{data.list}{A list of scRNA-seq count matrix.}

\item{gene.de}{A vector of ground-truth temporally DE genes.}
}
\value{
A list of precision, recall, and TNR of identified temporally DE genes on multiple datasets.
}
\description{
Infer temporally DE genes by Slingshot and GAM model on multiple datasets. Calculate corresponding precision, recall, and TNR of identified doublets.
}
\examples{
doublet.list <- FindDoublets(score.list = score.list, rate = 0.2)
data.removal.list <- RemoveDoublets.Method(count = count, label = label, doublet.list = doublet.list)
data.removal.list[['Clean Data']] <- list(count=count.clean, label=label.clean)
de.temp.result.all <- FindTempDE.All(data.list=data.removal.list, gene.de=gene.de)

}
