% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream.R
\name{FindTempDE}
\alias{FindTempDE}
\title{Infer temporally DE genes and calculate corresponding accuracy}
\usage{
FindTempDE(count, gene.de)
}
\arguments{
\item{count}{A scRNA-seq data matrix.}

\item{gene.de}{A vector of ground-truth temporally DE genes.}
}
\value{
A list of precision, recall, and TNR of identified temporally DE genes.
}
\description{
Infer temporally DE genes by Slingshot and GAM model on one dataset. Calculate precision, recall, and TNR of identified doublets.
}
\examples{
data.trajectory <- readRDS('.../sim_temporally_DE.rds')
count <- data.trajectory$count
gene.de <- data.trajectory$gene.de
de.temp.list <- FindTempDE(count=count, gene.de=gene.de)

}
