% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream.R
\name{FindTrajectory}
\alias{FindTrajectory}
\title{Conduct cell trajectory inference}
\usage{
FindTrajectory(count, label, title)
}
\arguments{
\item{count}{A scRNA-seq data matrix.}

\item{label}{A vector of doublet annotations (0/1).}

\item{title}{A title of the visualization.}
}
\value{
No return values.
}
\description{
Conduct cell trajectory inference by Slingshot on one dataset and draw the trajectory.
}
\examples{
data.trajectory <- readRDS('.../sim_psudotime_bifurcating.rds')
count <- data.trajectory$count
label <- data.trajectory$label
FindTrajectory(count=count, label=label, title='Contaminated Data')

}
