% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcessing.R
\name{ListToDataframe}
\alias{ListToDataframe}
\title{Transform a list to a dataframe}
\usage{
ListToDataframe(l, type)
}
\arguments{
\item{l}{A list of AUPRCs or AUROCs of different doublet-detection methods on multiple datasets.}

\item{type}{A charactor of "boxplot", "lineplot", "heatmap", "barplot", or "distributed" that refer to different format.}
}
\value{
A dataframe with approriate format for visualization.
}
\description{
Transform a list to a dataframe for visualization.
}
\examples{
result.auprc <- ListToDataframe(auprc.list.all, 'boxplot')
result.auroc <- ListToDataframe(auroc.list.all, 'boxplot')

}
