% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{Plot_Barplot}
\alias{Plot_Barplot}
\title{Barplot}
\usage{
Plot_Barplot(result, measurement)
}
\arguments{
\item{result}{A data frame with appropriate format for visualization.}

\item{measurement}{A title of visualization.}
}
\value{
No return values.
}
\description{
Plot Barplot for visualizing DE gene analysis.
}
\examples{
Plot_Barplot(table.DE.all[table.DE.all$measurement=='precision',], 'Precision')
Plot_Barplot(table.DE.all[table.DE.all$measurement=='recall',], 'Recall')
Plot_Barplot(table.DE.all[table.DE.all$measurement=='tnr',], 'TNR')

}
