% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{Plot_Barplot_temp}
\alias{Plot_Barplot_temp}
\title{The temporally DE genes version of Barplot}
\usage{
Plot_Barplot_temp(result, title)
}
\arguments{
\item{result}{A data frame with approriate format for visualization.}

\item{title}{A title of the visualization.}
}
\value{
No return values.
}
\description{
Plot barplot for visualizing temporally DE genes analysis.
}
\examples{
table.DE.temp <- ListToDataframe(l = de.temp.result.all, type = 'barplot')
Plot_Barplot_temp(result = table.DE.temp, title = 'Temporally DE Genes')

}
