% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{Plot_Boxplot}
\alias{Plot_Boxplot}
\title{Boxplot}
\usage{
Plot_Boxplot(result, measurement)
}
\arguments{
\item{result}{A dataframe with approriate format for visualization.}

\item{measurement}{A character to show the title of boxplot.}
}
\value{
No return value.
}
\description{
Plot a boxplot that shows the measurement distributions for different methods on multiple datasets.
Methods are sorted based on average performance.
}
\examples{
Plot_Boxplot(result = result.auprc, measurement = 'AUPRC')
Plot_Boxplot(result = result.auroc, measurement = 'AUROC')

}
