% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{Plot_Heatmap}
\alias{Plot_Heatmap}
\title{Heatmap}
\usage{
Plot_Heatmap(result, cluster)
}
\arguments{
\item{result}{A dataframe with approriate format for visualization.}

\item{cluster}{A number of true clusters in the dataset (numeric).}
}
\value{
No return values.
}
\description{
Plot Heatmap for visualizing cell clustering result.
}
\examples{
table.cluster <- ListToDataframe(l = result.cluster.all.rate, type = 'heatmap'); table.cluster
Plot_Heatmap(result = table.cluster, cluster = 4)
Plot_Heatmap(result = table.cluster, cluster = 6)
Plot_Heatmap(result = table.cluster, cluster = 8)

}
