% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{Plot_Lineplot}
\alias{Plot_Lineplot}
\title{Lineplot}
\usage{
Plot_Lineplot(result, type, measurement)
}
\arguments{
\item{result}{A dataframe with approriate format for visualization.}

\item{type}{A character to show the parameter of scRNA-seq data.}

\item{measurement}{A character to show the measurement of lineplot.}
}
\value{
No return values.
}
\description{
Plot a lineplot across different parameters of scRNA-seq data for multiple doublet-detection methods.
}
\examples{

result.auprc <- ListToDataframe(auprc.list.all, type = 'lineplot')
result.auroc <- ListToDataframe(auroc.list.all, type = 'lineplot')
Plot_Lineplot(result = result.auprc, type = 'Heterogeneity', measurement = 'AUPRC')
Plot_Lineplot(result = result.auroc, type = 'Doublet Rate', measurement = 'AUROC')

}
