% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcessing.R
\name{ReadData}
\alias{ReadData}
\title{Read rds datasets}
\usage{
ReadData(path = NULL)
}
\arguments{
\item{path}{A string of a user-specified directory.}
}
\value{
A list of data matrix and doublet annotations.
}
\description{
Read all rds files under current working directory or a user-specified directory.
}
\examples{
data.list <- ReadData(path = ".../real_datasets")

}
