% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindDoublets.R
\name{RemoveDoublets.All.Rate}
\alias{RemoveDoublets.All.Rate}
\title{Remove doublets for multiple methods on multiple datasets and doublet rates}
\usage{
RemoveDoublets.All.Rate(count.list, label.list, doublet.list.all.rate)
}
\arguments{
\item{count.list}{A list of scRNA-seq count matrices.}

\item{label.list}{A list doublet annoations (0/1).}

\item{doublet.list.all.rate}{A list of identified doublet indices of different methods on multiple datasets and doublet rates.}
}
\value{
A list of scRNA-seq count matrices and doublet annotations after removing doublets based on multiple doublet rates.
}
\description{
Remove doublets based on identified doublet indices for different doublet-detection methods on multiple datasets and doublet rates.
}
\examples{
doublet.list.all.rate <- FindDoublets.All.Rate(score.list.all = score.list.all, rates = seq(0.01, 0.25, 0.01))

}
