% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindDoublets.R
\name{RemoveDoublets.all}
\alias{RemoveDoublets.all}
\title{Remove doublets for multiple methods on multiple datasets}
\usage{
RemoveDoublets.all(count.list, label.list, doublet.list.all)
}
\arguments{
\item{count.list}{A list of scRNA-seq count matrices.}

\item{label.list}{A list doublet annoations (0/1).}

\item{doublet.list.all}{A list of identified doublet indices of different methods on multiple datasets.}
}
\value{
A list of scRNA-seq count matrices and doublet annotations after removing doublets.
}
\description{
Remove doublets based on identified doublet indices for different doublet-detection methods on multiple datasets.
}
\examples{
data.removal.all <- RemoveDoublets.All(count.list=count.list, label.list=label.list, doublet.list.all=doublet.list.all)

}
