% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindDoublets.R
\name{RemoveDoublets.Method}
\alias{RemoveDoublets.Method}
\title{Remove doublets for multiple methods on single dataset}
\usage{
RemoveDoublets.Method(count, label, doublet.list)
}
\arguments{
\item{count}{A scRNA-seq data matrix.}

\item{label}{A vector of cell type annotations.}

\item{doublet.list}{A list of identified doublet indices of different methods.}
}
\value{
A list of scRNA-seq data matrix and cell type annotations after removing doublets by different methods.
}
\description{
Remove doublets based on identified doublet indices for multiple methods on single dataset.
}
\examples{
data.removal.list <- RemoveDoublets.Method(count = data.de$count, label = data.de$label.cluster, doublet.list = doublet.list)

}
