% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindDoublets.R
\name{RemoveDoublets}
\alias{RemoveDoublets}
\title{Remove doublets for single method on single dataset}
\usage{
RemoveDoublets(count, label, doublets)
}
\arguments{
\item{count}{A scRNA-seq data matrix.}

\item{label}{A vector of cell type annotations.}

\item{doublets}{A vector of identified doublet indices.}
}
\value{
A list of scRNA-seq data matrix and cell type annotations after removing doublets.
}
\description{
Remove doublets based on identified doublet indices for single method on single dataset.
}
\examples{

}
