% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataProcessing.R
\name{SplitData}
\alias{SplitData}
\title{Split data into different batches}
\usage{
SplitData(count, label, batch)
}
\arguments{
\item{count}{A scRNA-seq data matrix.}

\item{label}{A vector of doublet annotations.}

\item{batch}{A batch number.}
}
\value{
A list of randomly split datasets
}
\description{
Split data into different batches.
}
\examples{

}
