% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FindDoublets.R
\name{FindDoublets}
\alias{FindDoublets}
\title{Obtain doublet scores and doublets}
\usage{
FindDoublets(counts, method, rate = NULL)
}
\arguments{
\item{rate}{}
}
\value{

}
\description{
Call different computational doublet-detection methods to obtain doublet scores and doublets.
}
